Add files via upload
The purpose of the paper is to analyse high risk countries by calculating deaths per cases. High risk countries are some of the highest in the following 4 categories, oldest population, smoking, asthma and cancer. The mentioned health concerns have been indicated by health professionals to comprise vulnerable groups of people and who are of higher risk when contracting Covid-19.